fix(endpoints): Handle lat/lng db queries correctly
